05.19.02
Defense: int32 -> float
피격시 무적 기본옵션 (0.5초)추가

+++ Source/AGSD/AGSDCharacter.h
@@ -196,7 +196,7 @@ public:
 	//캐릭터 스탯
 	int32 MaxHealth;        // 최대 체력
 	int32 CurrentHealth;    // 현제 체력
-	int32 Defense;          // 방어력
+	float Defense;          // 방어력
     float SpeedLevel;       // 이동속도
     int32 Barrier;           // 방어막
